requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
pycryptodome==3.19.0
python-socketio==5.10.0
python-engineio==4.8.0
Werkzeug==3.0.1
